stage names, stage & substage abbreviations in XML: https://github.com/metanorma/metanorma-iso/issues/342
